[WM] add authority create endpoint
This was helpful for me while debugging the UTF-8 encoding issues and fairly easy to add. No pressing need for it in production yet, and, in fact, it could be maliciously abused there since we have no auth on it.

So let's just leave in this in a branch for now.